Validate Login and find correct User
TDD is somewhat enjoyable.
When i have the energy to seek how to do stuff without trying them out
straight on the code.

Definitely worth it if you've already learned how to make the thing
you're making with TDD without TDD. Not my case tho...

Still a good experience. And the project is small enough not to get tiring.