Add Common Game Interface
The idea is to have every game instance live as an object of a specific
game class. The game class will contain all of the game mechanics and
will expose the interface necessary for a controller to let the user
interact with it.

That way the different thing in a game will be it's game class and
the view file for it. The controller logic will be shared.

The code is written according to what i tried in the prototype.

Well, i can't really make just a single controller action - it has to pass
an object to the view in it's original type.
But i can abstract the game logic in private method and have
the individual controller actions be around 5 lines each.